Move Holiday stuff to shared where feasible.
