+++ Content.Server/Holiday/Christmas/LimitedItemGiverSystem.cs
@@ -1,5 +1,7 @@
-﻿using Content.Server.Hands.Systems;
+using Content.Server.Hands.Systems;
 using Content.Server.Popups;
+using Content.Shared.Holiday;
+using Content.Shared.Holiday.Christmas;
 using Content.Shared.Interaction;
 using Content.Shared.Storage;
 using Robust.Shared.Player;
