+++ Content.Shared/Holiday/HolidayPrototype.cs
@@ -1,9 +1,9 @@
-using Content.Server.Holiday.Greet;
-using Content.Server.Holiday.Interfaces;
-using Content.Server.Holiday.ShouldCelebrate;
+using Content.Shared.Holiday.Greet;
+using Content.Shared.Holiday.Interfaces;
+using Content.Shared.Holiday.ShouldCelebrate;
 using Robust.Shared.Prototypes;
 
-namespace Content.Server.Holiday
+namespace Content.Shared.Holiday
 {
     [Prototype]
     public sealed partial class HolidayPrototype : IPrototype
