+++ Content.Shared/Holiday/Interfaces/IHolidayShouldCelebrate.cs
@@ -1,4 +1,4 @@
-namespace Content.Server.Holiday.Interfaces
+namespace Content.Shared.Holiday.Interfaces
 {
     [ImplicitDataDefinitionForInheritors]
     public partial interface IHolidayShouldCelebrate
